+++ cim-anonymiser/src/test/java/com/powsybl/cim/CimAnonymizerToolTest.java
@@ -1,0 +1,66 @@
+/**
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.cim;
+
+import com.google.common.io.ByteStreams;
+import com.powsybl.tools.AbstractToolTest;
+import com.powsybl.tools.CommandLineTools;
+import com.powsybl.tools.Tool;
+import net.java.truevfs.comp.zip.ZipEntry;
+import net.java.truevfs.comp.zip.ZipOutputStream;
+import org.junit.Test;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collections;
+
+import static org.junit.Assert.assertTrue;
+
+/**
+ * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
+ */
+public class CimAnonymizerToolTest extends AbstractToolTest {
+
+    private final CimAnonymizerTool tool = new CimAnonymizerTool();
+
+    @Override
+    protected Iterable<Tool> getTools() {
+        return Collections.singleton(tool);
+    }
+
+    @Override
+    public void assertCommand() {
+        assertCommand(tool.getCommand(), "anonymize-cim", 4, 3);
+        assertOption(tool.getCommand().getOptions(), "cim-zip-path", true, true);
+        assertOption(tool.getCommand().getOptions(), "output-dir", true, true);
+        assertOption(tool.getCommand().getOptions(), "dic-file", true, true);
+        assertOption(tool.getCommand().getOptions(), "skip-external-ref", false, false);
+    }
+
+    @Test
+    public void run() throws IOException {
+        Path workDir = fileSystem.getPath("work");
+        Path cimZipFile = workDir.resolve("sample.zip");
+        Path anonymizedCimFileDir = workDir.resolve("result");
+        Files.createDirectories(anonymizedCimFileDir);
+        Path dictionaryFile = workDir.resolve("dic.csv");
+        try (ZipOutputStream zos = new ZipOutputStream(Files.newOutputStream(cimZipFile))) {
+            zos.putNextEntry(new ZipEntry("sample_EQ.xml"));
+            zos.write(ByteStreams.toByteArray(getClass().getResourceAsStream("/sample_EQ.xml")));
+        }
+
+        assertCommand(new String[] {"anonymize-cim", "--cim-zip-path", cimZipFile.toString(),
+                                    "--output-dir", anonymizedCimFileDir.toString(),
+                                    "--dic-file", dictionaryFile.toString()},
+                CommandLineTools.COMMAND_OK_STATUS, "Anonymizing work/sample.zip", "");
+
+        assertTrue(Files.exists(anonymizedCimFileDir.resolve("sample.zip")));
+        assertTrue(Files.exists(dictionaryFile));
+        // anonymized content validation has already been done in other unit test
+    }
+}
