+++ cim-anonymiser/src/test/resources/sample.csv
@@ -1,0 +1,4 @@
+L1;A
+_BV_10;B
+Line 1;C
+10;D
