+++ cim-anonymiser/src/test/resources/sample_EQ.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:cim="http://iec.ch/TC57/2009/CIM-schema-cim14#">
+<cim:ACLineSegment rdf:ID="L1">
+<cim:Conductor.gch>0</cim:Conductor.gch>
+<cim:Conductor.x>0.001</cim:Conductor.x>
+<cim:Conductor.bch>0</cim:Conductor.bch>
+<cim:Conductor.r>0.0008</cim:Conductor.r>
+<cim:ConductingEquipment.BaseVoltage rdf:resource="#_BV_10"/>
+<cim:IdentifiedObject.name>Line 1</cim:IdentifiedObject.name>
+</cim:ACLineSegment>
+<cim:BaseVoltage rdf:ID="_BV_10">
+<cim:BaseVoltage.isDC>false</cim:BaseVoltage.isDC>
+<cim:BaseVoltage.nominalVoltage>1</cim:BaseVoltage.nominalVoltage>
+<cim:IdentifiedObject.name>10</cim:IdentifiedObject.name>
+</cim:BaseVoltage>
+</rdf:RDF>
